Switch region_dx in metarcld.fd to refer to projection grid spacing
Original code used the model grid spacing as the value for region_dx,
but that was because the map projection and model grid were identical.
For the new code, the map projection and model can have different grid
spacings, and for this program, the relevant one is the map projection
grid spacing. This grid spacing is now referred to as proj_dx instead of
region_dx.