Initial draft of map_proj_helper module code
Update code utilizes a new module for defining map projections for
larccld.fd, lightning.fd, and metarcld.fd. This code has been placed in
a separate branch because preliminary comparisons indicate that moving
the map projection definition to a separate module changes the output
(for reasons that I do not fully understand).