Add ability to dump cld_cover_3d from cloudanalysis.fd
Added a new namelist option for cloudanalysis.fd to dump the
cld_cover_3d field to the MPAS output. Updated code compiles and runs
without error, and output is identical to previous code versions when
this namelist option is turned off. It is expected that this new option
will help with debugging.